% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{adjCells}
\alias{adjCells}
\title{Get Adjacent Cell Coordinates}
\usage{
adjCells(r, xy)
}
\arguments{
\item{r}{A numeric vector of length 2 representing the resolution of the raster in x and y directions (e.g., \code{c(xres, yres)}).}

\item{xy}{A numeric vector of length 2 specifying the x and y coordinates of the center cell (e.g., \code{c(x, y)}).}
}
\value{
A matrix of 8 rows and 2 columns containing the coordinates of the adjacent cells.
}
\description{
Calculates the coordinates of 8 adjacent cells surrounding a given center cell.
}
\examples{
adjCells(c(10, 10), c(100, 100))
}
\keyword{internal}
