% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{adjRowCol}
\alias{adjRowCol}
\title{Get Row/Column Indices of Adjacent Cells}
\usage{
adjRowCol(rowcol)
}
\arguments{
\item{rowcol}{A numeric vector of length 2 representing the row and column index of the center cell.}
}
\value{
A matrix of 8 rows and 2 columns with the row and column indices of the adjacent cells.
}
\description{
Computes the row and column indices of the 8 adjacent cells surrounding a given central cell in a raster grid.
}
\examples{
adjRowCol(c(10, 10))
}
\keyword{internal}
