% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_to_raster.R
\name{cdfProb}
\alias{cdfProb}
\title{Convert Raster Values to Cumulative Distribution}
\usage{
cdfProb(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of empirical probabilities
}
\description{
Applies an empirical cumulative distribution function (ECDF) to raster values,
returning a raster where each cell reflects its percentile rank.
}
\examples{
\dontrun{
test <- cdfProb(c(1,2,3,4,5))
}
}
