% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{elevQuantile}
\alias{elevQuantile}
\title{Extract DEM elevations for a given quantile within a polygon}
\usage{
elevQuantile(dem, runout_ply, quant = 0.05, upper = TRUE)
}
\arguments{
\item{dem}{A \code{SpatRaster} object representing a digital elevation model (DEM).}

\item{runout_ply}{A \code{SpatVector} polygon (or compatible \verb{SpatialPolygons*} object) defining the area from which to extract elevation values.}

\item{quant}{A numeric value between 0 and 1 indicating the proportion (quantile) of elevation values to extract. Defaults to 0.05 (top/bottom 5\%).}

\item{upper}{Logical. If \code{TRUE}, the function selects the highest elevation percentile; if \code{FALSE}, it selects the lowest. Defaults to \code{TRUE}.}
}
\value{
A \code{SpatRaster} object where selected cells are marked with 1 and all others are set to \code{NA}.
}
\description{
Identifies and masks the top or bottom quantile of elevation values
within a given polygon overlaying a digital elevation model (DEM).
}
\details{
This function is useful for identifying areas of extreme elevation within a polygon, such as potential landslide release zones (upper) or deposition zones (lower), based on percentile thresholds.
}
\examples{
\dontrun{
library(terra)
dem <- rast("dem.tif")
runout_ply <- vect("runout_area.shp")
release_zone <- elevQuantile(dem, runout_ply, quant = 0.05, upper = TRUE)
}

}
