% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Calculate Euclidean Distance Between Two Points}
\usage{
euclideanDistance(p1, p2)
}
\arguments{
\item{p1}{A numeric vector of length 2 representing the first point's coordinates (x, y).}

\item{p2}{A numeric vector of length 2 representing the second point's coordinates (x, y).}
}
\value{
A numeric value representing the distance between the two points.
}
\description{
Computes the Euclidean distance between two 2D points.
}
\examples{
euclideanDistance(c(0, 0), c(3, 4))
}
