% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{makeSourceList}
\alias{makeSourceList}
\title{Create a list of xy coordinates from terra::xyFromCell coordinates}
\usage{
makeSourceList(source_xy)
}
\arguments{
\item{source_xy}{An \code{sf} object with POINT geometry or a
matrix with coordinate pairs (see \code{terra::xyFromCell})}
}
\description{
Create a list of xy coordinates from terra::xyFromCell coordinates
}
\examples{
\dontrun{
source_areas<- rast("Dev/Data/classified_w7filter_source_areas.tif")
source_areas <- crop(source_areas, dem)
sum(values(source_areas), na.rm = TRUE)

# Find cells where the value is 1
source_cells <- which(values(source_areas) == 1)

source_xy <- xyFromCell(source_areas, source_cells)

source_l <- makeSourceList(source_xy)
}
}
